goserbench: Split benchmarks into own file
This will be useful in the future to write a report generator.
